Move endpoint metadata export to JVM backend emit
